Add Reports back to RT 4.4
This commit adds the reports from 3.8 back to RT 4.4.
The reports are added under Tools >> Reports and
there are currently the 3 reports from 3.8.

The menus have been updated to 4.4 standard, and
the tabs have been updated to match the new search
link style.  Current GD based charts are working,
and the existing search and chart elements are used
throughout.